% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_gene.R
\name{annotation_gene}
\alias{annotation_gene}
\title{crawling NCBI gene annotation data in batches}
\usage{
annotation_gene(genelist)
}
\arguments{
\item{genelist}{a vector of gene SYMBOL}
}
\value{
get NCBI gene annotation data
}
\description{
crawling NCBI gene annotation data in batches
}
\examples{
sample<-head(annotation_gene(vertex_sample[,1]))
}
\author{
Ji Yang
}
