% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_sample.R
\docType{data}
\name{edge_sample}
\alias{edge_sample}
\title{sample data for edge file}
\format{
A dataframe with 2 columns, which are:
\describe{
\item{node1}{the end of edge}
\item{node2}{the other end of edge}
}
}
\usage{
edge_sample
}
\description{
a simple data of edge file that is  for simple testing and
running of examples
}
\keyword{datasets}
