% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetTopPara.R
\name{NetTopPara}
\alias{NetTopPara}
\title{calculate network topological parameters}
\usage{
NetTopPara(edge)
}
\arguments{
\item{edge}{the edge data of your network,such as edge_sample}
}
\value{
a dataframe with network topological parameters such as
degree,betweenness,closeness,eccentricity and eigenvector
}
\description{
calculate network topological parameters
}
\examples{
NetTopPara(edge_sample)
}
