% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDBinfGet.R
\name{PDBinfGet}
\alias{PDBinfGet}
\title{Get all available PDB_ID information of the genelist you input}
\usage{
PDBinfGet(genelist,
          id.fromType = "hgnc_symbol",
          id.biomart  = ENSEMBL_MART_ENSEMBL,
          id.dataset  = "hsapiens_gene_ensembl",
          id.toType   = "uniprot_gn_id"
}
\arguments{
\item{genelist}{a vector of gene(s)}

\item{id.fromType}{input id type(must be same as type of input genelist),we set 4 frequently-used types here,
which are "hgnc_symbol","entrezgene_id","ensembl_gene_id","uniprot_gn_id"}

\item{id.biomart}{biomart types when id needs to be translated,
see also \code{\link{biomart::getBM}}}

\item{id.dataset}{biomart dataset choose,
see also \code{\link{biomart::getBM}}}

\item{id.toType}{output id type}
}
\value{
properties of PDB_ID,such as method of Protein crystal analysis,
resolution,chains,structure length
}
\description{
we acquire information of PDB_ID in virtue of the api of
EMBL protein(https://www.ebi.ac.uk/proteins/api/proteins/),you just need
to input a vector of genelist in your hand.
}
\examples{
#input type is "hgnc_symbol":
PDBinfGet(genelist    = vertex_sample[1:5,1],
          id.fromType = "hgnc_symbol",
          id.biomart  = ENSEMBL_MART_ENSEMBL,
          id.dataset  = "hsapiens_gene_ensembl",
          id.toType   = "uniprot_gn_id")
#input type is "uniprot_gn_id":
PDBinfGet(genelist    = "P21580",
          id.fromType = "uniprot_gn_id",
          id.biomart  = ENSEMBL_MART_ENSEMBL,
          id.dataset  = "hsapiens_gene_ensembl",
          id.toType   = "uniprot_gn_id")

}
