% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psv2n.R
\name{psv2n}
\alias{psv2n}
\title{calculate score by PS-V2N methods}
\usage{
psv2n(vertex, edge)
}
\arguments{
\item{vertex}{the vertex data of your network,such as vertex_sample}

\item{edge}{the edge data of your network,such as edge_sample}
}
\value{
the PS-V2N score of your group A
}
\description{
1) ready for one dataframe(such as vertex_sample, first
column is gene name, and the second colname is the type
identification code 0, 1, and 2 (0 means groupA, 1 maens groupB, 2 means share genes);
2) Get your edge dataframe ready(such as edge_sample)
}
\examples{
psv2n(vertex_sample, edge_sample)
}
\author{
Ji Yang
}
