% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertex_input.R
\name{vertex_input}
\alias{vertex_input}
\title{Generate the input file of vertex for next psv2n computation}
\usage{
vertex_input(groupA, groupB)
}
\arguments{
\item{groupA}{a list of genes}

\item{groupB}{a list of target genes that you want to acquire \code{\link{psv2n}} score}
}
\value{
the result of two kinds of gene list named groupA and groupB
}
\description{
We have two kinds of gene list named groupA and groupB,
we need integrate these files into one.
}
\examples{
vertex_sample
}
