% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sample.R
\docType{data}
\name{vertex_sample}
\alias{vertex_sample}
\title{sample data for vertex file}
\format{
A dataframe with 2 columns, which are:
\describe{
\item{node}{the node of network}
\item{aorb}{the type of node, 0 means groupB, 1 maens groupA, 2 means share genes}
}
}
\usage{
vertex_sample
}
\description{
a simple data of vertex file that is  for simple testing and
running of examples
}
\keyword{datasets}
