% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOplot.R
\name{GOplot}
\alias{GOplot}
\title{plot GO plot with your interisting genelist}
\usage{
GOplot(
  genelist,
  display.number = 10,
  enrich.pvalue = 0.05,
  CPCOLS = c("#8DA1CB", "#FD8D62", "#66C3A5"),
  labelsize = 8,
  label.title = "The Most Enriched GO Terms",
  xlab.title = "GO Term"
)
}
\arguments{
\item{genelist}{a vector of gene SYMBOL}

\item{display.number}{number of display terms,respectively}
}
\value{
get the GO plot
}
\description{
plot GO plot with your interisting genelist
}
\examples{
GOplot(vertex_sample[,1])
}
\author{
Ji Yang
}
