% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGplot.R
\name{KEGGplot}
\alias{KEGGplot}
\title{plot KEGG plot with your interisting genelist}
\usage{
KEGGplot(
  genelist,
  enrich.pvalue = 0.05,
  low.color = "green",
  high.color = "red",
  labs.x = "Pvalue",
  labs.y = "Pathway Name",
  labs.title = "Pathway Enrichment"
)
}
\arguments{
\item{genelist}{a vector of gene SYMBOL}

\item{labs.title}{}
}
\description{
plot KEGG plot with your interisting genelist
}
\examples{
KEGGplot(vertex_sample[,1])
}
